Added a weighted random option picker
Added the ability to pick a random number between 0 and total weight number. The game then uses this number to determine which option was selected, weighted proportionally to each options weight variable number. All of this was done in the game-manager object, which I added to the room to make it work.